Add colmap step
